Wfprev 183 a (#804)

+++ server/wfprev-api/src/main/java/ca/bc/gov/nrs/wfprev/data/entities/EvaluationCriteriaSectionSummaryEntity.java
@@ -29,6 +29,7 @@ import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
@@ -68,7 +69,7 @@ public class EvaluationCriteriaSectionSummaryEntity implements Serializable {
 
     @NotNull
     @Column(name = "filter_section_score", columnDefinition = "numeric(5) default '0'", nullable = false)
-    private Integer filterSectionScore;
+    private BigDecimal filterSectionScore;
 
 
     @Column(name = "filter_section_comment", length = 4000)
